# perf: disable progress bar by default