--- a/website/docs/en/config/dev/progress-bar.mdx
+++ b/website/docs/en/config/dev/progress-bar.mdx
@@ -10,15 +10,17 @@ type ProgressBar =
     };
 ```
 
-- **Default:** `process.env.NODE_ENV === 'production'`
+- **Default:** `false`
 
-Whether to display progress bar during compilation.
+Whether to render progress bars to display the build progress.
 
 By default, Rsbuild only enables the progress bar when production building.
 
-- **Example:** Enable the progress bar.
+## Example
 
-```js
+Enable the progress bar:
+
+```ts title="rsbuild.config.ts"
 export default {
   dev: {
     progressBar: true,
@@ -26,9 +28,9 @@ export default {
 };
 ```
 
-- **Example:** Disable the progress bar.
+Disable the progress bar:
 
-```js
+```ts title="rsbuild.config.ts"
 export default {
   dev: {
     progressBar: false,
@@ -36,15 +38,53 @@ export default {
 };
 ```
 
-- **Example:** Modify the progress bar `id`
+Only enable the progress bar in development mode:
+
+```ts title="rsbuild.config.ts"
+const isDev = process.env.NODE_ENV === 'development';
+
+export default {
+  dev: {
+    progressBar: isDev,
+  },
+};
+```
+
+## Customize the progress bar `id`
 
 To modify the text displayed on the left side of the progress bar, set the `id` option:
 
-```ts
+```ts title="rsbuild.config.ts"
 export default {
   dev: {
     progressBar: {
       id: 'Some Text',
+    },
+  },
+};
+```
+
+When using [environments](/guide/advanced/environments), you can set different progress bar `id` for different environments:
+
+```ts title="rsbuild.config.ts"
+export default {
+  environments: {
+    web: {
+      dev: {
+        progressBar: {
+          id: 'Web',
+        },
+      },
+    },
+    node: {
+      output: {
+        target: 'node',
+      },
+      dev: {
+        progressBar: {
+          id: 'Node',
+        },
+      },
     },
   },
 };