--- a/website/docs/en/config/dev/progress-bar.mdx
+++ b/website/docs/en/config/dev/progress-bar.mdx
@@ -14,7 +14,9 @@ type ProgressBar =
 
 Whether to render progress bars to display the build progress.
 
-By default, Rsbuild only enables the progress bar when production building.
+:::tip
+In `@rsbuild/core < 1.4.0`, the progress bar is enabled by default in production mode.
+:::
 
 ## Example