… packages/core/tests/__snapshots__/default.test.ts.snap