## Summary

Disable the progress bar for production build by default.

- In small and medium applications, Rspack builds fast enough that we usually don't need to render the progress bar.
- In large applications, rendering the progress bar will increase the build time by 1% to 10%, so we prefer to disable the progress bar by default.
- I am also not satisfied with the current progress bar style, it is not elegant enough and may be interrupted by the log output during the build process.

## Performance

Tested in https://github.com/rspack-contrib/build-tools-performance.

### 20k modules

```bash
hyperfine --warmup 2 --runs 4  'CASE=large PROGRESS=true node --run build:rsbuild'  'CASE=large node --run build:rsbuild'

Benchmark 1: CASE=large PROGRESS=true node --run build:rsbuild
  Time (mean ± σ):      1.681 s ±  0.008 s    [User: 6.291 s, System: 2.584 s]
  Range (min … max):    1.670 s …  1.689 s    4 runs
 
Benchmark 2: CASE=large node --run build:rsbuild
  Time (mean ± σ):      1.543 s ±  0.038 s    [User: 6.142 s, System: 2.486 s]
  Range (min … max):    1.507 s …  1.591 s    4 runs
 
Summary
  CASE=large node --run build:rsbuild ran
    1.09 ± 0.03 times faster than CASE=large PROGRESS=true node --run build:rsbuild
```

### 2.5k modules

```bash
hyperfine --warmup 2 --runs 4  'CASE=small PROGRESS=true node --run build:rsbuild'  'CASE=small node --run build:rsbuild'

Benchmark 1: CASE=small PROGRESS=true node --run build:rsbuild
  Time (mean ± σ):     434.2 ms ±   4.1 ms    [User: 1064.5 ms, System: 449.2 ms]
  Range (min … max):   430.5 ms … 438.8 ms    4 runs
 
Benchmark 2: CASE=small node --run build:rsbuild
  Time (mean ± σ):     428.8 ms ±   3.0 ms    [User: 1060.8 ms, System: 452.3 ms]
  Range (min … max):   426.1 ms … 431.5 ms    4 runs
 
Summary
  CASE=small node --run build:rsbuild ran
    1.01 ± 0.01 times faster than CASE=small PROGRESS=true node --run build:rsbuild
```

## Checklist

<!--- Check and mark with an "x" -->

- [x] Tests updated (or not required).
- [x] Documentation updated (or not required).